ai-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sentence-transformers==2.5.0
pydantic==2.6.1
pydantic-settings==2.1.0
numpy==1.26.4
torch==2.2.0
redis==5.0.1
httpx==0.26.0
python-multipart==0.0.9
pymupdf==1.23.22
python-docx==1.1.0
spacy==3.7.4
